uzhavan
initial commit

+++ api/loan_issue_files/delete_doc_info.php
@@ -6,15 +6,17 @@ $result = 0;
 $qry = $pdo->query("SELECT upload FROM `document_info` WHERE id='$id'");
 if($qry->rowCount()>0){
     $row = $qry->fetch();
-    unlink("../../uploads/loan_issue/doc_info/" . $row['upload']);
-    
+    if ($row['upload']) {
+        unlink("../../uploads/loan_issue/doc_info/" . $row['upload']);
+        }
+    }
     $qry = $pdo->query("DELETE FROM `document_info` WHERE id='$id'");
     if ($qry) {
         $result = 1;
     }else{
         $result = 2;
     }
-}
+
 
 $pdo = null; //Close connection.
