+++ api/loan_issue_files/delete_mortgage_info.php
@@ -6,15 +6,17 @@ $result = 0;
 $qry = $pdo->query("SELECT upload FROM `mortgage_info` WHERE id='$id'");
 if($qry->rowCount()>0){
     $row = $qry->fetch();
+    if ($row['upload']) {
     unlink("../../uploads/loan_issue/mortgage_info/" . $row['upload']);
-    
+    }
+}
     $qry = $pdo->query("DELETE FROM `mortgage_info` WHERE id='$id'");
     if ($qry) {
         $result = 1;
     }else{
         $result = 2;
     }
-}
+
 
 $pdo = null; //Close connection.
